feat: you-goal-b08-retired-surface-enforcement-006 - Gate closeout with lint and verify-pr
